test: migrate PopupBase test
